+++ COM/test/Wall.h
@@ -43,8 +43,8 @@ public:
 
 	// IUnknown functions
 	virtual HRESULT __stdcall QueryInterface(const IID& interfaceid, void** objectInterface);
-	virtual ULONG __stdcall AddRef();
-	virtual ULONG __stdcall Release();
+	virtual UINT32 __stdcall AddRef();
+	virtual UINT32 __stdcall Release();
 	
 	virtual HRESULT readWallBSTR(BSTR* text);
 	virtual HRESULT writeBSTROnWall(BSTR text);
